Solo Box Collision Fully Solved
I designed an algorithm for solo box collision.
It is a fully functional algorithm without the bugs of the previous approach. It was determined via mathematical induction using geometry and basic algebra.
Its efficiency could certainly be improved, but as a prototype and baseline implementation, it is effective.

The next task is to solve Duo Box Collision where two pushable boxes collide which