Duo Box Collision Fully Solved
I designed an algorithm for duo box collision. This algorithm for the box collider allows two entities which are marked to "solve their collisions" to push eachother based upon their current velocity.

This means that if an Entity which moves at 2 pixels per frame collides with an Entity which moves at 3 pixels per frame, the faster entity will push the other one back at a speed of 1 pixel per frame!

It is a fully functional algorithm, determined via mathematical induction using geometry, basic algebra, and reference of the solo box collision resolution.
Its efficiency can certainly be improved, but as a prototype and baseline implementation, it is effective.

The next task is to make serialization & deserialization of scenes and set up a pipeline for them to be written to & read from files.